tests/requirements.txt
pytest==8.2.0
pytest-timeout==2.3.1
requests==2.32.3
psycopg2-binary==2.9.9
confluent-kafka==2.4.0
